+++ CHANGELOG.md
@@ -19,7 +19,7 @@
 * Updated deps:
 
 ```clj
-[com.cognitect/transit-clj "0.8.293"] is available but we use "0.8.290"
+[com.cognitect/transit-clj "0.8.295"] is available but we use "0.8.290"
 [com.fasterxml.jackson.core/jackson-databind "2.8.5"] is available but we use "2.8.4"
 ```