+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## 0.2.0-SNAPSHOT
 
+* The new `muuntaja.json` JSON encoder & decoder.
+  * directly on top of [Jackson](https://github.com/FasterXML/jackson)
+  * explicit mappings instead of protocol extensions
+  * encoding is 2.5 - 5.5x faster than Cheshire
+  * decoding is 30%+ faster than Cheshire
+
 * **BREAKING**: move and rename http-negotiation keys from top level to `:http` in options:
   * `:extract-content-type-fn` =>  `:extract-content-type`
   * `:extract-accept-charset-fn` => `:extract-accept-charset`
@@ -9,7 +15,6 @@
 * **BREAKING**: `muuntaja.options` namespace is thrown away.
   * new helpers in `muuntaja.core`: `transform-formats` & `select-formats`
   * `muuntaja.options/default-options-with-format-regexps` can be copy-pasted from below:
-* default-options support all JVM registered charsets (instead of just `utf-8`)
 
 ```clj
 (def default-options-with-format-regexps
@@ -21,6 +26,9 @@
       (assoc-in [:formats "application/transit+json" :matches] #"^application/(vnd.+)?(x-)?transit\+json$")
       (assoc-in [:formats "application/transit+msgpack" :matches] #"^application/(vnd.+)?(x-)?transit\+msgpack$"))
 ```
+
+* default-options support all JVM registered charsets (instead of just `utf-8`)
+* re-organized namespaces & code: formats now in separate namespaces
 
 * Updated deps: